Launch surveydown Studio

This function serves two purposes:

It generates a new survey project from a template

It launches a Shiny app to preview and edit the survey

Usage

sd_studio(path = getwd(), template = "default")

Arguments

path

A character string specifying the directory where the survey template should be created. Defaults to the current working directory.

template

A character string specifying the template to use. Default is "default" which uses the built-in package template. See surveydown::sd_create_survey() for other available templates.

Value

No return value, called for its side effects of creating a survey project and launching a Shiny app.

Examples

if (interactive()) {
  # Launch studio with default template
  sd_studio()

  # Launch studio with a specific template and path
  sd_studio(path = "my_survey", template = "question_types")
}